small test changes, correct networkclient as per the code review to throw exceptions

+++ SmartTabsTest/src/ch/epfl/sweng/smartTabs/test/TimeTest.java
@@ -19,7 +19,7 @@ public class TimeTest extends AndroidTestCase{
 	}
 	
 	public void testDuration(){
-		assertEquals((double)5, time.getDuration());
+		assertEquals(5, time.getDuration());
 	}
 	
 	public void testTernary(){
